sepparate 'entry' (unit) test and 'scenario' test for cssc
