+++ src/acrcssc/azext_acrcssc/tests/latest/test_cssc_scenario.py
@@ -6,53 +6,38 @@
 import unittest
 from unittest import mock
 from unittest.mock import MagicMock
-from azure.cli.testsdk import ScenarioTest
+from azure.cli.testsdk import ScenarioTest, ResourceGroupPreparer
 from azext_acrcssc.cssc import create_acrcssc, update_acrcssc, delete_acrcssc, show_acrcssc, cancel_runs, list_scan_status
 
 class AcrcsscScenarioTest(ScenarioTest):
-    def __init__(self, method_name, config_file=None, recording_name=None, recording_processors=None, replay_processors=None, recording_patches=None, replay_patches=None, random_config_dir=False):
-        super().__init__(method_name, config_file, recording_name, recording_processors, replay_processors, recording_patches, replay_patches, random_config_dir)
-        self.cmd = mock.MagicMock()
-        self.cmd.cli_ctx = MagicMock()
-        self.registry = MagicMock()
-        self.registry.name = "mockregistry"
-        self.registry.id = f"/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/mockrg/providers/Microsoft.ContainerRegistry/registries/{self.registry.name}"
-        self.config = MagicMock()
 
-    @mock.patch("azext_acrcssc.cssc._perform_continuous_patch_operation")
-    def test_create_acrcssc(self, mock_perform_continuous_patch_operation):
-        create_acrcssc(self.cmd, "mockrg", self.registry.name, "continuouspatchv1", "mockconfig", "1d", False, False)
-        mock_perform_continuous_patch_operation.assert_called_once_with(self.cmd, "mockrg", self.registry.name, "mockconfig", "1d", False, False, is_create=True)
+    @ResourceGroupPreparer()
+    def test_acrcssc_workflow(self, resource_group):
+        self.kwargs.update({
+            'registry_name': self.create_random_name('clireg', 20),
+            'workflow_type': 'scan',
+            'config': 'path/to/config.json',
+            'schedule': '7d'
+        })
 
-    @mock.patch("azext_acrcssc.cssc._perform_continuous_patch_operation")
-    def test_update_acrcssc(self, mock_perform_continuous_patch_operation):
-        update_acrcssc(self.cmd, "mockrg", self.registry.name, "continuouspatchv1", "mockconfig", "1d", False, False)
-        mock_perform_continuous_patch_operation.assert_called_once_with(self.cmd, "mockrg", self.registry.name, "mockconfig", "1d", False, False, is_create=False)
+        self.cmd('acr create -n {registry_name} -g {rg} --sku Standard',
+                 checks=[self.check('name', '{registry_name}'),
+                         self.check('provisioningState', 'Succeeded')])
 
-    @mock.patch("azext_acrcssc.cssc.delete_continuous_patch_v1")
-    @mock.patch("azext_acrcssc.cssc.cf_acr_registries")
-    @mock.patch("azure.cli.core.util.user_confirmation")
-    def test_delete_acrcssc(self, mock_user_confirmation, mock_cf_acr_registries, mock_delete_continuous_patch_v1):
-        mock_cf_acr_registries.return_value.get.return_value = self.registry
-        delete_acrcssc(self.cmd, "mockrg", self.registry.name, "continuouspatchv1")
-        mock_user_confirmation.assert_called_once()
-        mock_delete_continuous_patch_v1.assert_called_once_with(self.cmd, self.registry, False)
+        self.cmd('acr supply-chain workflow create -r {registry_name} -t {workflow_type} --config {config} --schedule {schedule}',
+                 checks=[self.check('workflowType', '{workflow_type}'),
+                         self.check('provisioningState', 'Succeeded')])
 
-    @mock.patch("azext_acrcssc.cssc.list_continuous_patch_v1")
-    @mock.patch("azext_acrcssc.cssc.cf_acr_registries")
-    def test_show_acrcssc(self, mock_cf_acr_registries, mock_list_continuous_patch_v1):
-        mock_cf_acr_registries.return_value.get.return_value = self.registry
-        show_acrcssc(self.cmd, "mockrg", self.registry.name, "continuouspatchv1")
-        mock_list_continuous_patch_v1.assert_called_once_with(self.cmd, self.registry)
+        self.cmd('acr supply-chain workflow list -r {registry_name}',
+                 checks=[self.check('[0].workflowType', '{workflow_type}')])
 
-    @mock.patch("azext_acrcssc.cssc.cancel_continuous_patch_runs")
-    def test_cancel_runs(self, mock_cancel_continuous_patch_runs):
-        cancel_runs(self.cmd, "mockrg", self.registry.name, "continuouspatchv1")
-        mock_cancel_continuous_patch_runs.assert_called_once_with(self.cmd, "mockrg", self.registry.name)
+        self.cmd('acr supply-chain workflow show -r {registry_name} -t {workflow_type}',
+                 checks=[self.check('workflowType', '{workflow_type}')])
 
-    @mock.patch("azext_acrcssc.cssc.track_scan_progress")
-    @mock.patch("azext_acrcssc.cssc.cf_acr_registries")
-    def test_list_scan_status(self, mock_cf_acr_registries, mock_track_scan_progress):
-        mock_cf_acr_registries.return_value.get.return_value = self.registry
-        list_scan_status(self.cmd, self.registry.name, "mockrg", "mockstatus", "continuouspatchv1")
-        mock_track_scan_progress.assert_called_once_with(self.cmd, "mockrg", self.registry, "mockstatus")
+        self.cmd('acr supply-chain workflow update -r {registry_name} -t {workflow_type} --schedule 14d',
+                 checks=[self.check('schedule', '14d')])
+
+        self.cmd('acr supply-chain workflow delete -r {registry_name} -t {workflow_type}',
+                 checks=[self.check('provisioningState', 'Succeeded')])
+
+        self.cmd('acr delete -n {registry_name} -g {rg} -y')
